Fix NSME which is prevents new lines from being added
 #KT-13030 Fixed

The exception was thrown because of non-existent method in StringUtil.
It happened because we use shrinked version of it from kotlin-compiler
but method called from plugin code which could use full version. To
avoid this, we replace that piece of code which cause the exception,
namely we remove call 'createToStringFunction' from toString() method